initial commit
Project and structure created
The skeleton of the Products Table page (without logic) is created: two buttons and a table
Added a window to the password field (without logic)
Implemented logic to show/hide the password on the Login page
Made the skeleton of the Products Preview page (product cards, without the ability to click on them)
On the LogIn page was made:
Sending a token request, receiving a token, storing a token in cookies (frontend + backend)
Validation was done:
Fields must not be empty. Notification: if the login or password is incorrect.
When opening the Product Table page, products were retrieved from the server and filled in the table
If the login is successful and the token is in the cookie, it redirects the user to the ProductsTable page. If the token is deleted from the cookie and the page is reloaded, we are redirected to the Login page wherever we are.
On the ProductTable page, when clicking on the ‘Preview’ button, we created a redirect to the ‘Product Preview’

Translated with DeepL.com (free version)